'betterstats:s2c_bss' packet
This feature prevents the client's access to the 'betterstats' network protocol until the server confirms it has the 'betterstats' mod installed on its end. This is done to further protect the client from servers and so the client has an indicator as to whether or not the server has 'betterstats' installed.